django_blog/requirements.txt
asgiref==3.9.2
Django==5.2.7
django-filter==22.1
djangorestframework==3.16.1
dotenv==0.9.9
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
mysqlclient==2.2.7
pillow==11.3.0
django-taggit
